Refactor: 여러 부분 리팩토링

+++ src/main/java/nbc/chillguys/nebulazone/application/user/dto/response/AdminUserResponse.java
@@ -19,7 +19,7 @@ public record AdminUserResponse(
 	String email,
 	String phone,
 	String nickname,
-	int point,
+	long point,
 	OAuthType oAuthType,
 	String oAuthId,
 	Set<AddressResponse> addresses,
